requirements.txt
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.3
openai==1.37.1
Werkzeug==3.0.3
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
blinker==1.8.2
tqdm==4.66.4